Ticket 49021 - Automatic thread tuning
Bug Description:  We currently always set 30 threads on systems no matter their
size. We should tune the threads more appropriately out of the box.

Fix Description:  Add a thread autotuning. We start at 16 threads, and ramp up,
reducing the rate off addition as we grow to a cap of 512 threads.

https://fedorahosted.org/389/ticket/49021

Author: wibrown

Review by: nhosoi, mreynolds (Thanks)

+++ ldap/servers/slapd/proto-slap.h
@@ -472,7 +472,7 @@ char *config_get_rootpwstoragescheme(void);
 char *config_get_localuser(void);
 char *config_get_workingdir(void);
 char *config_get_encryptionalias(void);
-int config_get_threadnumber(void);
+long config_get_threadnumber(void);
 int config_get_maxthreadsperconn(void);
 int config_get_maxdescriptors(void);
 int config_get_reservedescriptors(void);
